# src/main/resources/public/docs/openapi.yml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: Pulmonis API
...
paths:
  /public/news:
    get:
      tags:
        - Public News
      summary: Get all news for visitor
      parameters:
        - in: path
          name: page
          schema:
            type: integer
            default: 0
            description: Parameter for specifing page of news to be retrieved. If not provided, page=0 will be returned. Default size of page is 16 and elements sorted by update time.
            example: 3
        - in: header
          name: find-latest
          required: false
          schema:
            type: boolean
            default: "false"
          description: Flag to find 4 latest news with provided language which are published.
        - $ref: "#/components/parameters/lang"
      description: Public endpoint for getting all news in a specific language without any authentication.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationListingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        412:
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /public/news/{newsId}:
    get:
      tags:
        - Public News
      summary: Get published news for visitor
      parameters:
        - $ref: "#/components/parameters/newsId"
        - $ref: "#/components/parameters/lang"
      description: Public endpoint for getting a news piece in a specific language without any authentication.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /public/articles:
    get:
      tags:
        - Public Article
      summary: Get all articles for visitor
      parameters:
        - in: path
          name: page
          schema:
            type: integer
            default: 0
            description: Parameter for specifing page of articles to be retrieved. If not provided, page=0 will be returned. Default size of page is 16 and elements sorted by update time.
            example: 3
        - $ref: "#/components/parameters/lang"
      description: Public endpoint for getting all articles in a specific language without any authentication.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationListingResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        412:
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
  /public/articles/{articleId}:
    get:
      tags:
        - Public Article
      summary: Get published article for visitor
      parameters:
        - $ref: "#/components/parameters/articleId"
        - $ref: "#/components/parameters/lang"
      description: Public endpoint for getting an article in a specific language without any authentication.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
  /news/{newsId}:
    get:
      tags:
        - News
      summary: One specific news by its id.
      parameters:
        - in: path
          name: newsId
          required: true
          schema:
            type: integer
            minimum: 1
          description: The news ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullPublicationResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
  /articles/{articleId}:
    get:
      tags:
        - Article
      summary: One specific article by its id.
      parameters:
        - in: path
          name: articleId
          required: true
          schema:
            type: integer
            minimum: 1
          description: The article's ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullPublicationResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
        '500':
          description: Server Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsuccessfulResponse"
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    lang:
      in: query
      name: lang
      schema:
        type: string
        enum: [ et, en, ru ]
        example: et
      required: true
      description: Publication language
    articleId:
      in: path
      name: articleId
      required: true
      schema:
        type: integer
        minimum: 1
      description: The article's ID
    newsId:
      in: path
      name: newsId
      required: true
      schema:
        type: integer
        minimum: 1
      description: The news ID
  schemas:
...
    FullPublicationResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/FullPublicationData"
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
    PublicationResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/PublicationData"
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
...
    PublicationListingResponse:
      type: object
      properties:
        data:
          properties:
            totalItems:
              type: integer
              example: 1
              description: Total items returned.
            publications:
              type: array
              items:
                $ref: "#/components/schemas/PublicationData"
            totalPages:
              type: integer
              example: 1
              description: Total pages.
            currentPage:
              type: integer
              example: 0
              description: Currently selected page.
        errors:
          $ref: "#/components/schemas/EmptyError"
        status:
          $ref: "#/components/schemas/SuccessfulStatus"
...
    FullPublicationData:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Publication ID
          example: 7
        titleEn:
          type: string
          nullable: true
          example: Test title
          description: Publication title in English
        titleEt:
          type: string
          nullable: true
          example: null
          description: Publication title in Estonian
        titleRu:
          type: string
          nullable: true
          example: null
          description: Publication title in Russian
        descriptionEn:
          type: string
          nullable: true
          example: Test description
          description: Publication description in English
        descriptionEt:
          type: string
          nullable: true
          example: null
          description: Publication description in Estonian
        descriptionRu:
          type: string
          nullable: true
          example: null
          description: Publication description in Russian
        bodyEn:
          type: string
          nullable: true
          example: Test body
          description: Publication body in English
        bodyEt:
          type: string
          nullable: true
          example: null
          description: Publication body in Estonian
        bodyRu:
          type: string
          nullable: true
          example: null
          description: Publication body in Russian
        publishedEt:
          type: boolean
          example: true
          description: Is the publication published in Estonian
        publishedEn:
          type: boolean
          example: false
          description: Is the publication published in English
        publishedRu:
          type: boolean
          example: false
          description: Is the publication published in Russian
        category:
          type: string
          example: TestCategory
          description: Category of the publication
        publishTime:
          type: date-time
          description: Last publication time in format of yyyy-mm-ddThh:mm:ss
          example: 2015-12-03T10:15:35
    PublicationData:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Publication ID
          example: 7
        title:
          type: string
          nullable: true
          example: Test title
          description: Publication title
        description:
          type: string
          nullable: true
          example: Test description
          description: Publication description
        body:
          type: string
          nullable: true
          example: Test body
          description: Publication body
        published:
          type: boolean
          example: true
          description: Is the publication published in parameter specified language
        category:
          type: string
          example: TestCategory
          description: Category of the publication
        publishTime:
          type: date-time
          description: Last publication time in format of yyyy-mm-ddThh:mm:ss
          example: 2015-12-03T10:15:35
...
    EmptyError:
      type: array
      nullable: true
      description: List of errors returned by API.
      example: null
    SuccessfulStatus:
      type: integer
      example: 200
      default: 200
      description: Successful response from API.
    UnsuccessfulStatus:
      type: integer
      example: 400
      description: Successful status response from API. Can vary from 400 to 500.
    UnsuccessfulResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
          example: null
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
        status:
          $ref: "#/components/schemas/UnsuccessfulStatus"
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: Error message returned from API.
          nullable: true
          example: "string"
        code:
          type: string
          description: Error code returned from API.
          example: "string"
security:
  - bearerAuth: [ ]